+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Additional Changes:
 * Memory leaks removed from unit tests
 * Upgrade to jasmine 3.0.0
 * Refactored tests and CI with Saucelabs
+* WEB-1680: messageStatusRenderer and dateRenderer properties now used on initial load as well as new messages
 
 
 ## 1.0.0-pre2.2